% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_rmd}
\alias{config_rmd}
\title{Config rmd}
\usage{
config_rmd(config_yml = "_config.yml")
}
\arguments{
\item{config_yml}{the path to your config.yml}
}
\description{
A function to be run to add the Rmd regex to the exclude in config.yml
}
