% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_post.R
\name{create_post}
\alias{create_post}
\title{Create a post}
\usage{
create_post(title, dir = "_posts", url = title, date = Sys.Date())
}
\arguments{
\item{title}{post title}

\item{dir}{output directory (default to _posts)}

\item{url}{the output url}

\item{date}{the date of the post}
}
\value{
creates a post
}
\description{
Create a post
}
